requirements.txt
fastapi
uvicorn
python-multipart
numpy
pillow
python-dotenv
torch
torchvision
ultralytics
piexif
